refactor: move type definitions to data_types.py
